feat: add traceable functionality & side effects to prototypes
Inspired by 'ngrx' module for Angular project which helps trace application's state in different parts of the application and act accordingly, this module provides functions to make any prototype a state traceable.
This means that any instance of a class that holds a state can be traceable and have consumers to it's state updates using Observable pattern.
Additionally, in the prototype level it also possible to define side effects that run automatically on instance creation, perform actions which might change the instance state.
It is also possible to define static effects that run regardless a particular state.